+++ include/igraph_operators.h
@@ -64,7 +64,7 @@ IGRAPH_EXPORT igraph_error_t igraph_contract_vertices(igraph_t *graph,
                                            const igraph_attribute_combination_t
                                      *vertex_comb);
 IGRAPH_EXPORT igraph_error_t igraph_permute_vertices(const igraph_t *graph, igraph_t *res,
-                                          const igraph_vector_t *permutation);
+                                          const igraph_vector_int_t *permutation);
 IGRAPH_EXPORT igraph_error_t igraph_connect_neighborhood(igraph_t *graph, igraph_integer_t order,
                                               igraph_neimode_t mode);
 IGRAPH_EXPORT igraph_error_t igraph_rewire(igraph_t *graph,